requirements.txt
datasets==3.1.0
numpy==1.24.4
matplotlib==3.7.5
Pillow==10.4.0
torch==2.2.2
fsspec==2024.9.0
flask==3.1.0